General: Change No. 1 Standardized the tick process to do proper tick alocation. Change No. 2 Added a code block to launch shop titans if it is not open. Change No. 3 Moved the interactions with customers to its own script.
Crafting:
Change No. 4
Cleaned up the crafting code blocks so that the kings caprice & standard crafter use the same common code block
Bounty:

Change No. 5
Added a task collector

Market:
Change No. 6
Added a script to collect and relist items on the

Quests:
Change No. 7
Updated the Tower of Titans script to work with the new patch(there was a small change to the ui that prevented the script from detecting the event)
Change No. 8
Added a subscription free version of the quest collector